+++ .eslintrc.js
@@ -1,5 +1,12 @@
 module.exports = {
     extends: [
         '@nextcloud'
-    ]
+    ],
+    settings: {
+        'import/resolver': {
+            'node': {
+                'extensions': ['.js', '.ts', 'vue']
+            }
+        }
+    }
 };